PlainValueTemplate class made partial with encoding detail separated
